lint fix
- Replace non-null assertion (!) with optional chaining (?.) in suggestion handler tests
- Move isInViewport from BlockEditorUtils to BlockEditorPureUtils to break circular dependency:
  hashtagSuggestion → HashList → BlockEditorUtils → BlockEditorExtensionsClassBase → hashtagSuggestion
