+++ openmetadata-ui/src/main/resources/ui/src/components/BlockEditor/Extensions/hashtag/hashtagSuggestion.test.ts
@@ -109,7 +109,7 @@ describe('hashtagSuggestion getReferenceClientRect null safety', () => {
       handler.onStart(makeSuggestionProps(() => null) as SuggestionProps);
 
       const [, options] = mockTippy.mock.calls[0];
-      const result = options.getReferenceClientRect!();
+      const result = options.getReferenceClientRect?.();
 
       expect(result).toBeInstanceOf(DOMRect);
     });
@@ -121,7 +121,7 @@ describe('hashtagSuggestion getReferenceClientRect null safety', () => {
       handler.onStart(makeSuggestionProps(() => rect) as SuggestionProps);
 
       const [, options] = mockTippy.mock.calls[0];
-      const result = options.getReferenceClientRect!();
+      const result = options.getReferenceClientRect?.();
 
       expect(result).toBe(rect);
     });
@@ -133,7 +133,7 @@ describe('hashtagSuggestion getReferenceClientRect null safety', () => {
 
       const [, options] = mockTippy.mock.calls[0];
 
-      expect(() => options.getReferenceClientRect!()).not.toThrow();
+      expect(() => options.getReferenceClientRect?.()).not.toThrow();
     });
   });
