+++ openmetadata-ui/src/main/resources/ui/src/components/BlockEditor/Extensions/mention/mentionSuggestions.test.ts
@@ -90,7 +90,7 @@ describe('mentionSuggestion getReferenceClientRect null safety', () => {
       handler.onStart(makeSuggestionProps(() => null) as SuggestionProps);
 
       const [, options] = mockTippy.mock.calls[0];
-      const result = options.getReferenceClientRect!();
+      const result = options.getReferenceClientRect?.();
 
       expect(result).toBeInstanceOf(DOMRect);
     });
@@ -102,7 +102,7 @@ describe('mentionSuggestion getReferenceClientRect null safety', () => {
       handler.onStart(makeSuggestionProps(() => rect) as SuggestionProps);
 
       const [, options] = mockTippy.mock.calls[0];
-      const result = options.getReferenceClientRect!();
+      const result = options.getReferenceClientRect?.();
 
       expect(result).toBe(rect);
     });
@@ -114,7 +114,7 @@ describe('mentionSuggestion getReferenceClientRect null safety', () => {
 
       const [, options] = mockTippy.mock.calls[0];
 
-      expect(() => options.getReferenceClientRect!()).not.toThrow();
+      expect(() => options.getReferenceClientRect?.()).not.toThrow();
     });
   });
