+++ openmetadata-ui/src/main/resources/ui/src/components/BlockEditor/Extensions/slash-command/renderItems.test.ts
@@ -103,7 +103,7 @@ describe('renderItems getReferenceClientRect null safety', () => {
       handler.onStart(makeSuggestionProps(() => null) as SuggestionProps);
 
       const [, options] = mockTippy.mock.calls[0];
-      const result = options.getReferenceClientRect!();
+      const result = options.getReferenceClientRect?.();
 
       expect(result).toBeInstanceOf(DOMRect);
     });
@@ -115,7 +115,7 @@ describe('renderItems getReferenceClientRect null safety', () => {
       handler.onStart(makeSuggestionProps(() => rect) as SuggestionProps);
 
       const [, options] = mockTippy.mock.calls[0];
-      const result = options.getReferenceClientRect!();
+      const result = options.getReferenceClientRect?.();
 
       expect(result).toBe(rect);
     });
@@ -127,7 +127,7 @@ describe('renderItems getReferenceClientRect null safety', () => {
 
       const [, options] = mockTippy.mock.calls[0];
 
-      expect(() => options.getReferenceClientRect!()).not.toThrow();
+      expect(() => options.getReferenceClientRect?.()).not.toThrow();
     });
   });
