+++ openmetadata-ui/src/main/resources/ui/src/utils/BlockEditorPureUtils.ts
@@ -201,3 +201,13 @@ export const stripPendingUploadNodes = (html: string): string => {
 
   return doc.body.innerHTML;
 };
+
+export const isInViewport = (ele: HTMLElement, container: HTMLElement) => {
+  const eleTop = ele.offsetTop;
+  const eleBottom = eleTop + ele.clientHeight;
+
+  const containerTop = container.scrollTop;
+  const containerBottom = containerTop + container.clientHeight;
+
+  return eleTop >= containerTop && eleBottom <= containerBottom;
+};
