+++ openmetadata-ui/src/main/resources/ui/src/utils/BlockEditorUtils.ts
@@ -36,16 +36,6 @@ export const getSelectedText = (state: EditorState) => {
   return text;
 };
 
-export const isInViewport = (ele: HTMLElement, container: HTMLElement) => {
-  const eleTop = ele.offsetTop;
-  const eleBottom = eleTop + ele.clientHeight;
-
-  const containerTop = container.scrollTop;
-  const containerBottom = containerTop + container.clientHeight;
-
-  // The element is fully visible in the container
-  return eleTop >= containerTop && eleBottom <= containerBottom;
-};
 
 // Unique marker prefix used to temporarily replace entity links during HTML serialization
 // This avoids HTML encoding of < and > characters in entity links
